0410_commit

+++ src/Practice1/Project6.java
@@ -5,13 +5,18 @@ public class Project6 {
 	public static void main(String[] args) {
 		// TODO 自動生成されたメソッド・スタブ
 		// 変数宣言
-		int x;
+		double x;
 		// 値代入
 		x = 7;
+		
+		//xの値を3倍にする
 		x = 3 * x;
+		//表示
 		System.out.println(x);
 
+		//xの値を1/2する
 		x = x / 2;
+		//表示
 		System.out.println(x);
 
 	}
